# backend/docs/swagger.yaml
basePath: /v1
definitions:
  domain.AnswerResultData:
    properties:
      is_correct:
        type: boolean
      played_word:
        type: string
    type: object
...
  domain.BadgeItem:
    properties:
      code:
        type: string
      criteria_type:
        type: string
      criteria_value:
        type: integer
      description_vi:
        type: string
      earned_at:
        type: string
      icon_url:
        type: string
      id:
        type: string
      name_vi:
        type: string
    type: object
  domain.BadgeListData:
    properties:
      earned:
        items:
          $ref: '#/definitions/domain.BadgeItem'
        type: array
      locked:
        items:
          $ref: '#/definitions/domain.BadgeItem'
        type: array
    type: object
  domain.ChartsData:
    properties:
      period:
        type: string
      points:
        items:
          $ref: '#/definitions/domain.SnapshotPoint'
        type: array
    type: object
...
  domain.DrillStatusData:
    properties:
      correct_count:
        type: integer
      drill_id:
        type: string
      hearts_left:
        type: integer
      status:
        type: string
      total_items:
        type: integer
    type: object
...
  domain.ListenDrillData:
    properties:
      drill_id:
        type: string
      hearts_left:
        type: integer
      pairs:
        items:
          $ref: '#/definitions/domain.ListenPairItem'
        type: array
      total_items:
        type: integer
    type: object
  domain.ListenPairItem:
    properties:
      audio_a:
        type: string
      audio_b:
        type: string
      pair_id:
        type: string
      word_a:
        type: string
      word_b:
        type: string
    type: object
...
  domain.ProgressOverviewData:
    properties:
      current_streak:
        type: integer
      last_active_date:
        type: string
      longest_streak:
        type: integer
      total_sessions:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  domain.SnapshotPoint:
    properties:
      date:
        type: string
      score:
        type: number
    type: object
...
  handler.answerRequest:
    properties:
      chosen_word:
        type: string
      pair_id:
        type: string
    required:
    - chosen_word
    - pair_id
    type: object
...
  handler.startDrillRequest:
    properties:
      count:
        maximum: 20
        minimum: 5
        type: integer
    type: object
...
  service.StreakDTO:
    properties:
      current_streak:
        type: integer
      last_active_date:
        type: string
      longest_streak:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@phonara.app
    name: Phonara Team
  description: |-
    Backend API cho ứng dụng luyện phát âm tiếng Anh Phonara.
    Bao gồm: Authentication, Practice Sessions, Error Profile Engine,
    Content, Shadowing, Minimal Pairs, Subscription, Exam (Speechace), Daily Challenge.
  license:
    name: Proprietary
  title: Phonara API
  version: "1.0"
paths:
...
  /badges:
    get:
      description: Trả earned[] và locked[] với criteria và progress tới badge kế
        tiếp (FR-STREAK-03, S20).
      produces:
      - application/json
      responses:
        "200":
          description: earned and locked badges
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.BadgeListData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách badges
      tags:
      - Progress
...
  /minimal-pairs/listen/{drill_id}:
    get:
      parameters:
      - description: Drill UUID
        in: path
        name: drill_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: total_items, correct_count, hearts_left, status
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.DrillStatusData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Trạng thái listen drill
      tags:
      - MinimalPairs
  /minimal-pairs/listen/{drill_id}/answer:
    post:
      consumes:
      - application/json
      description: Chấm đúng/sai, trừ heart nếu sai, cập nhật pair_mastery.listen
        (FR-MP-02).
      parameters:
      - description: Drill UUID
        in: path
        name: drill_id
        required: true
        type: string
      - description: Answer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.answerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: is_correct + played_word
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AnswerResultData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Drill or pair not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Nộp đáp án listen drill
      tags:
      - MinimalPairs
  /minimal-pairs/listen/start:
    post:
      consumes:
      - application/json
      description: Tạo phiên nghe phân biệt âm với N cặp ngẫu nhiên, hearts=3 (FR-MP-02,
        S16).
      parameters:
      - description: Drill config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.startDrillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Drill created with pair list
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ListenDrillData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Bắt đầu listen drill
      tags:
      - MinimalPairs
  /progress/charts:
    get:
      description: Đọc mastery_snapshots để trả điểm theo thời gian cho sparkline
        (S19/S28).
      parameters:
      - description: week hoặc month
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chart data points
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChartsData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Score trend chart
      tags:
      - Progress
  /progress/overview:
    get:
      description: Streak hiện tại, longest_streak, tổng số session (S20).
      produces:
      - application/json
      responses:
        "200":
          description: Overview data
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProgressOverviewData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Progress overview
      tags:
      - Progress
...
  /streak/check-in:
    post:
      description: Ghi nhận hoạt động hôm nay theo timezone user. Tăng current_streak
        nếu liên tiếp, reset về 1 nếu gián đoạn, cập nhật longest_streak (BR-STREAK-01).
      produces:
      - application/json
      responses:
        "200":
          description: Updated streak
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.StreakDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Streak check-in hàng ngày
      tags:
      - Progress
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"